Rename `Element.measure` as `Element.measureReference` for two reasons:
* Reflect the object type, which is `MeasureReference` and not the also existing `Measure`.
* Make room for a `measure` property which returns the "real" `Measure` object.

The rational for direct association to `Measure` in `Element` is because Java interfaces
can execute code, which is different than XML or JSON documents where data are static.
Implementers are free to do themselves the work of fetching full `Measure` description
from a measure registry or catalogue, which is easier for users. This is optional;
implementers can provide only a `MeasureReference` (as ISO 19157) if they want.

https://github.com/opengeospatial/geoapi/issues/75